requirements.txt
pytest==7.4.3
requests==2.31.0
pydantic==2.5.3
selenium==4.16.0
Faker==22.5.1
sgqlc==16.3
python-dotenv==1.0.1
pymongo==4.6.2